프로토콜선언위치 변경
