--- JSONParser/JSONParser/JSONPrintable.swift
@@ -1,14 +1,0 @@
-//
-//  JSONPrintable.swift
-//  JSONParser
-//
-//  Created by moon on 2018. 5. 2..
-//  Copyright © 2018년 JK. All rights reserved.
-//
-
-import Foundation
-
-protocol JSONPrintable {
-    func totalDataCountDescription() -> String
-    func countDataDescription() -> String
-}
