+++ JSONParser/JSONParser/OutputView.swift
@@ -8,6 +8,11 @@
 
 import Foundation
 
+protocol JSONPrintable {
+    func totalDataCountDescription() -> String
+    func countDataDescription() -> String
+}
+
 struct OutputView {
     // 총 4개의 객체 데이터 중에 문자열 2개, 숫자 1개, 부울 1개가 포함되어 있습니다.
     // 총 2개의 배열 데이터 중에 객체 2개가 포함되어 있습니다.
